btf: move readAndInflateTypes
Move the implementation of readAndInflateTypes into a new file. No
other changes intended.
